# [hail][bugfix] Resolve performance bug in variant_qc