--- a/hail/python/hail/ir/ir.py
+++ b/hail/python/hail/ir/ir.py
@@ -2269,6 +2269,20 @@ class Literal(IR):
         self._type = self._typ
 
 
+class LiftMeOut(IR):
+    @typecheck_method(child=IR)
+    def __init__(self, child):
+        super().__init__(child)
+        self.child = child
+
+    def copy(self, child):
+        return LiftMeOut(child)
+
+    def _compute_type(self, env, agg_env):
+        self.child._compute_type(env, agg_env)
+        self._type = self.child.typ
+
+
 class Join(IR):
     _idx = 0