--- a/hail/src/main/scala/is/hail/expr/ir/Interpret.scala
+++ b/hail/src/main/scala/is/hail/expr/ir/Interpret.scala
@@ -674,6 +674,12 @@ object Interpret {
         wrapped.get(0)
       case x: ReadPartition =>
         fatal(s"cannot interpret ${ Pretty(x) }")
+      case LiftMeOut(child) =>
+        val (rt, makeFunction) = Compile[Long](ctx, MakeTuple.ordered(FastSeq(child)), None, false)
+        Region.scoped { r =>
+          SafeRow.read(rt, r, makeFunction(0, r)(r)).asInstanceOf[Row](0)
+        }
+
     }
   }
 }