We were reading the column values per row. This was bad.

Introduce the LiftMeOut node that is evaluated in the
`InterpretNonCompilable` lowering pass.

The benchmarked pipelines are still exponential in the number of
iterations of QC, which I will fix in a forthcoming PR.

Before:

```
2020-01-29 12:36:38,135: INFO: [1/1] Running variant_and_sample_qc_nested_with_filters...
2020-01-29 12:38:31,070: INFO: burn in: 112.93s
2020-01-29 12:40:17,863: INFO: run 1: 106.79s
2020-01-29 12:42:01,471: INFO: run 2: 103.61s
2020-01-29 12:43:47,330: INFO: run 3: 105.86s
```

After:
```
2020-01-29 12:30:18,619: INFO: [1/1] Running variant_and_sample_qc_nested_with_filters...
2020-01-29 12:31:24,969: INFO: burn in: 66.34s
2020-01-29 12:32:24,797: INFO: run 1: 59.82s
2020-01-29 12:33:22,849: INFO: run 2: 58.05s
2020-01-29 12:34:20,667: INFO: run 3: 57.81s
```

SSDs are fast.